creation du context pour l'api weather
